+++ cilicili-back/src/main/java/com/cilicili/entity/User.java
@@ -6,7 +6,7 @@ import lombok.Data;
 @Data
 @TableName("user")
 public class User {
-    private int id;
+    private Long id;
     private String username;
     private String password;
     private String nickname;
@@ -15,7 +15,7 @@ public class User {
     private String signature;
     private String email;
     private String phone;
-    private byte status;
+    private Byte status;
     private String created_at;
     private String updated_at;
 }
